--- monitor/dockerfile
@@ -1,19 +1,0 @@
-FROM python:3.9
-
-RUN apt-get update && apt-get install -y --no-install-recommends
-RUN pip install --upgrade pip
-
-# Install necessary libraries
-RUN pip install mysql-connector-python
-RUN pip install slack_sdk
-
-# Copy the code to the container
-COPY monitor.py /app/monitor.py
-COPY db_handler.py /app/db_handler.py
-
-# Set working directory
-WORKDIR /app
-
-# Run the Python script
-#ENTRYPOINT ["python", "monitor.py"]
-ENTRYPOINT ["/bin/bash"]
